+++ rust/dbn/src/decode/dbn/sync.rs
@@ -36,12 +36,13 @@ where
     /// Creates a new DBN [`Decoder`] from `reader`.
     ///
     /// # Errors
-    /// This function will return an error if it is unable to parse the metadata in `reader`.
+    /// This function will return an error if it is unable to parse the metadata in
+    /// `reader` or the input is encoded in a newer version of DBN.
     pub fn new(mut reader: R) -> crate::Result<Self> {
         let metadata = MetadataDecoder::new(&mut reader).decode()?;
         Ok(Self {
+            decoder: RecordDecoder::with_version(reader, metadata.version)?,
             metadata,
-            decoder: RecordDecoder::new(reader),
         })
     }
 
@@ -173,6 +174,8 @@ pub struct RecordDecoder<R>
 where
     R: io::Read,
 {
+    /// For future use with reading different DBN versions.
+    _version: u8,
     reader: R,
     buffer: Vec<u8>,
 }
@@ -184,10 +187,28 @@ where
     /// Creates a new `RecordDecoder` that will decode from `reader`.
     pub fn new(reader: R) -> Self {
         Self {
+            _version: DBN_VERSION,
             reader,
             // `buffer` should have capacity for reading `length`
             buffer: vec![0],
         }
+    }
+
+    /// Creates a new `RecordDecoder` that will decode from `reader`
+    /// with the specified DBN version.
+    ///
+    /// # Errors
+    /// This function will return an error if the `version` exceeds the supported version.
+    pub fn with_version(reader: R, version: u8) -> crate::Result<Self> {
+        if version > DBN_VERSION {
+            return Err(crate::Error::decode(format!("Can't decode newer version of DBN. Decoder version is {DBN_VERSION}, input version is {version}")));
+        }
+        Ok(Self {
+            _version: version,
+            reader,
+            // `buffer` should have capacity for reading `length`
+            buffer: vec![0],
+        })
     }
 
     /// Returns a mutable reference to the inner reader.